nike-instore-checker/requirements.txt
discord.py==2.3.0
requests==2.31.0